Powerlaw_Fit
Use of the Powerlaw package in Python to fit the Species and Genera data following several distributions; exponential, powerl_law, log_normal. Among these, It is the truncated power law that is the best fit to our data. Now it would be nice to compare the fit of this truncated power_law to the Zipf and Zipf-Mandelbrot laws