Zipf Modelisation
Modelisation of our data with the Zipf Law, using the same methodology we used for Zipf-Mandelbrot: does not work, I have the same issues. The modelisation with the log-log Monte-Carlo methods works decently  on the other hand: species, generas, and some of the generas taken individually seem to follow a Zipf-Mandelbrot Law; the others, not so much. Maybe a Truncated Power Law is more adapted to them